fixed cycled dependency between instruments and notation modules
