+++ src/instruments/instrumentsmodule.h
@@ -33,7 +33,6 @@ public:
     void resolveImports() override;
     void registerResources() override;
     void registerUiTypes() override;
-    void onInit(const framework::IApplication::RunMode& mode) override;
 };
 }
