+++ src/instruments/view/staffsettingsmodel.cpp
@@ -69,7 +69,7 @@ QString StaffSettingsModel::staffType() const
 
 void StaffSettingsModel::setStaffType(int type)
 {
-    auto type_ = static_cast<notation::StaffType>(type);
+    auto type_ = static_cast<StaffType>(type);
 
     if (m_type == type_ || !parts()) {
         return;
