+++ src/notation/CMakeLists.txt
@@ -55,7 +55,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/inotationplayback.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationelements.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationparts.h
-
+    ${CMAKE_CURRENT_LIST_DIR}/iinstrumentsrepository.h
+    ${CMAKE_CURRENT_LIST_DIR}/iselectinstrumentscenario.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/notationuiactions.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/notationuiactions.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/igetscore.h
@@ -118,6 +119,10 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/inotationmidievents.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/notationmidievents.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/notationmidievents.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsconverter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsconverter.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsrepository.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsrepository.h
     ${CMAKE_CURRENT_LIST_DIR}/view/inotationcontextmenu.h
     ${CMAKE_CURRENT_LIST_DIR}/view/notationpaintview.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/notationpaintview.h
@@ -178,7 +183,6 @@ set(MODULE_LINK
     qzip
     ${Z_LIB}
     engraving
-    instruments
     commonscene
     uicomponents
     ui
