+++ src/notation/inotationconfiguration.h
@@ -144,6 +144,18 @@ public:
 
     virtual void setTemplateModeEnalbed(bool enabled) = 0;
     virtual void setTestModeEnabled(bool enabled) = 0;
+
+    virtual io::paths instrumentListPaths() const = 0;
+    virtual async::Notification instrumentListPathsChanged() const = 0;
+
+    virtual io::paths userInstrumentListPaths() const = 0;
+    virtual void setUserInstrumentListPaths(const io::paths& paths) = 0;
+
+    virtual io::paths scoreOrderListPaths() const = 0;
+    virtual async::Notification scoreOrderListPathsChanged() const = 0;
+
+    virtual io::paths userScoreOrderListPaths() const = 0;
+    virtual void setUserScoreOrderListPaths(const io::paths& paths) = 0;
 };
 }
