+++ src/notation/inotationparts.h
@@ -23,7 +23,6 @@
 #define MU_NOTATION_INOTATIONPARTS_H
 
 #include "notationtypes.h"
-#include "instruments/instrumentstypes.h"
 #include "async/notification.h"
 #include "async/channel.h"
 #include "async/notifylist.h"
@@ -39,14 +38,14 @@ public:
     virtual ~INotationParts() = default;
 
     virtual async::NotifyList<const Part*> partList() const = 0;
-    virtual async::NotifyList<instruments::Instrument> instrumentList(const ID& partId) const = 0;
+    virtual async::NotifyList<Instrument> instrumentList(const ID& partId) const = 0;
     virtual async::NotifyList<const Staff*> staffList(const ID& partId, const ID& instrumentId) const = 0;
 
     virtual ValCh<bool> canChangeInstrumentVisibility(const ID& instrumentId, const ID& fromPartId) const = 0;
     virtual bool voiceVisible(int voiceIndex) const = 0;
 
-    virtual void setParts(const instruments::PartInstrumentList& instruments) = 0;
-    virtual void setScoreOrder(const instruments::ScoreOrder& order) = 0;
+    virtual void setParts(const PartInstrumentList& instruments) = 0;
+    virtual void setScoreOrder(const ScoreOrder& order) = 0;
     virtual void setPartVisible(const ID& partId, bool visible) = 0;
     virtual void setInstrumentVisible(const ID& instrumentId, const ID& fromPartId, bool visible) = 0;
     virtual void setStaffVisible(const ID& staffId, bool visible) = 0;
@@ -54,7 +53,7 @@ public:
     virtual void setVoiceVisible(const ID& staffId, int voiceIndex, bool visible) = 0;
     virtual void setPartName(const ID& partId, const QString& name) = 0;
     virtual void setPartSharpFlat(const ID& partId, const SharpFlat& sharpFlat) = 0;
-    virtual void setPartTransposition(const ID& partId, const instruments::Interval& transpose) = 0;
+    virtual void setPartTransposition(const ID& partId, const Interval& transpose) = 0;
     virtual void setInstrumentName(const ID& instrumentId, const ID& fromPartId, const QString& name) = 0;
     virtual void setInstrumentAbbreviature(const ID& instrumentId, const ID& fromPartId, const QString& abbreviature) = 0;
     virtual void setStaffType(const ID& staffId, StaffType type) = 0;
@@ -77,12 +76,12 @@ public:
                                  const ID& destinationInstrumentId, InsertMode mode = InsertMode::Before) = 0;
     virtual void moveStaves(const IDList& sourceStavesIds, const ID& destinationStaffId, InsertMode mode = InsertMode::Before) = 0;
 
-    virtual void appendDoublingInstrument(const instruments::Instrument& instrument, const ID& destinationPartId) = 0;
+    virtual void appendDoublingInstrument(const Instrument& instrument, const ID& destinationPartId) = 0;
     virtual void appendStaff(Staff* staff, const ID& destinationPartId) = 0;
 
     virtual void cloneStaff(const ID& sourceStaffId, const ID& destinationStaffId) = 0;
 
-    virtual void replaceInstrument(const ID& instrumentId, const ID& fromPartId, const instruments::Instrument& newInstrument) = 0;
+    virtual void replaceInstrument(const ID& instrumentId, const ID& fromPartId, const Instrument& newInstrument) = 0;
 
     virtual async::Notification partsChanged() const = 0;
 };
