+++ src/notation/internal/notation.h
@@ -51,7 +51,7 @@ public:
     Meta metaInfo() const override;
     void setMetaInfo(const Meta& meta) override;
 
-    instruments::ScoreOrder scoreOrder() const override;
+    ScoreOrder scoreOrder() const override;
 
     INotationPtr clone() const override;
