+++ src/notation/internal/notationconfiguration.h
@@ -148,7 +148,30 @@ public:
     void setTemplateModeEnalbed(bool enabled) override;
     void setTestModeEnabled(bool enabled) override;
 
+    io::paths instrumentListPaths() const override;
+    async::Notification instrumentListPathsChanged() const override;
+
+    io::paths userInstrumentListPaths() const override;
+    void setUserInstrumentListPaths(const io::paths& paths) override;
+
+    io::paths scoreOrderListPaths() const override;
+    async::Notification scoreOrderListPathsChanged() const override;
+
+    io::paths userScoreOrderListPaths() const override;
+    void setUserScoreOrderListPaths(const io::paths& paths) override;
+
 private:
+    io::path firstInstrumentListPath() const;
+    void setFirstInstrumentListPath(const io::path& path);
+
+    io::path secondInstrumentListPath() const;
+    void setSecondInstrumentListPath(const io::path& path);
+
+    io::path firstScoreOrderListPath() const;
+    void setFirstScoreOrderListPath(const io::path& path);
+
+    io::path secondScoreOrderListPath() const;
+    void setSecondScoreOrderListPath(const io::path& path);
 
     async::Notification m_backgroundChanged;
     async::Notification m_foregroundChanged;
@@ -156,6 +179,8 @@ private:
     async::Channel<framework::Orientation> m_canvasOrientationChanged;
     async::Channel<io::path> m_userStylesPathChanged;
     async::Channel<int> m_selectionColorChanged;
+    async::Notification m_instrumentListPathsChanged;
+    async::Notification m_scoreOrderListPathsChanged;
 };
 }
