+++ src/notation/notationmodule.cpp
@@ -38,6 +38,7 @@
 #include "internal/mscznotationreader.h"
 #include "internal/msczmetareader.h"
 #include "internal/positionswriter.h"
+#include "internal/instrumentsrepository.h"
 
 #include "view/notationpaintview.h"
 #include "view/notationswitchlistmodel.h"
@@ -65,6 +66,8 @@
 #include "view/notationcontextmenu.h"
 #include "view/internal/undoredomodel.h"
 
+#include "diagnostics/idiagnosticspathsregister.h"
+
 using namespace mu::notation;
 using namespace mu::modularity;
 using namespace mu::ui;
@@ -75,6 +78,7 @@ static std::shared_ptr<NotationConfiguration> s_configuration = std::make_shared
 static std::shared_ptr<NotationActionController> s_actionController = std::make_shared<NotationActionController>();
 static std::shared_ptr<NotationUiActions> s_notationUiActions = std::make_shared<NotationUiActions>(s_actionController);
 static std::shared_ptr<MidiInputController> s_midiInputController = std::make_shared<MidiInputController>();
+static std::shared_ptr<InstrumentsRepository> s_instrumentsRepository = std::make_shared<InstrumentsRepository>();
 
 static void notationscene_init_qrc()
 {
@@ -90,6 +94,7 @@ void NotationModule::registerExports()
 {
     ioc()->registerExport<INotationCreator>(moduleName(), new NotationCreator());
     ioc()->registerExport<INotationConfiguration>(moduleName(), s_configuration);
+    ioc()->registerExport<IInstrumentsRepository>(moduleName(), s_instrumentsRepository);
     ioc()->registerExport<IMsczMetaReader>(moduleName(), new MsczMetaReader());
     ioc()->registerExport<INotationContextMenu>(moduleName(), new NotationContextMenu());
 
@@ -186,9 +191,33 @@ void NotationModule::registerUiTypes()
 void NotationModule::onInit(const framework::IApplication::RunMode&)
 {
     s_configuration->init();
+    s_instrumentsRepository->init();
     s_actionController->init();
     s_notationUiActions->init();
     s_midiInputController->init();
 
     Notation::init();
+
+    auto pr = modularity::ioc()->resolve<diagnostics::IDiagnosticsPathsRegister>(moduleName());
+    if (pr) {
+        io::paths instrPaths = s_configuration->instrumentListPaths();
+        for (const io::path& p : instrPaths) {
+            pr->reg("instruments", p);
+        }
+
+        io::paths uinstrPaths = s_configuration->userInstrumentListPaths();
+        for (const io::path& p : uinstrPaths) {
+            pr->reg("user instruments", p);
+        }
+
+        io::paths scoreOrderPaths = s_configuration->scoreOrderListPaths();
+        for (const io::path& p : scoreOrderPaths) {
+            pr->reg("scoreOrder", p);
+        }
+
+        io::paths uscoreOrderPaths = s_configuration->userScoreOrderListPaths();
+        for (const io::path& p : uscoreOrderPaths) {
+            pr->reg("user scoreOrder", p);
+        }
+    }
 }
