+++ src/notation/notationtypes.h
@@ -27,6 +27,7 @@
 
 #include "io/path.h"
 #include "translation.h"
+#include "midi/midievent.h"
 
 #include "libmscore/element.h"
 #include "libmscore/page.h"
@@ -52,8 +53,6 @@
 #include "libmscore/realizedharmony.h"
 #include "libmscore/instrument.h"
 
-#include "instruments/instrumentstypes.h"
-
 namespace mu::notation {
 using Page = Ms::Page;
 using Element = Ms::Element;
@@ -64,7 +63,7 @@ using DurationType = Ms::TDuration::DurationType;
 using Duration = Ms::TDuration;
 using SelectType = Ms::SelectType;
 using Pad = Ms::Pad;
-using ViewMode = Ms::LayoutMode;  // Accomodate inconsistent convention from v3
+using ViewMode = Ms::LayoutMode;
 using PitchMode = Ms::UpDownMode;
 using StyleId = Ms::Sid;
 using SymbolId = Ms::SymId;
@@ -73,7 +72,6 @@ using KeyMode = Ms::KeyMode;
 using TimeSigType = Ms::TimeSigType;
 using Part = Ms::Part;
 using Staff = Ms::Staff;
-using StaffType = Ms::StaffTypes;
 using NoteHead = Ms::NoteHead;
 using SharpFlat = Ms::PreferSharpFlat;
 using TransposeMode = Ms::TransposeMode;
@@ -100,8 +98,24 @@ using TupletBracketType = Ms::TupletBracketType;
 using GraceNoteType = Ms::NoteType;
 using BeamMode = Ms::Beam::Mode;
 using LayoutBreakType = Ms::LayoutBreak::Type;
+using Interval = Ms::Interval;
+using Drumset = Ms::Drumset;
+using StringData = Ms::StringData;
+using Clef = Ms::Clef;
+using ClefType = Ms::ClefType;
+using ClefTypeList = Ms::ClefTypeList;
+using BracketType = Ms::BracketType;
+using StaffGroup = Ms::StaffGroup;
+using StaffType = Ms::StaffTypes;
+using StaffTypePreset = Ms::StaffType;
+using StaffName = Ms::StaffName;
+using StaffNameList = Ms::StaffNameList;
+using MidiArticulation = Ms::MidiArticulation;
+using Trait = Ms::Trait;
+using TraitType = Ms::TraitType;
 using InstrumentChannel = Ms::Channel;
 
+using InstrumentChannelList = QList<InstrumentChannel>;
 using PageList = std::vector<const Page*>;
 using StaffList = QList<const Staff*>;
 using PartList = QList<const Part*>;
@@ -282,6 +296,177 @@ struct Tempo
     }
 };
 
+static constexpr int MAX_STAVES  = 4;
+
+struct ClefPair
+{
+    ClefType concertClef = ClefType::G;
+    ClefType transposingClef = ClefType::G;
+};
+
+struct PitchRange
+{
+    int min = 0;
+    int max = 0;
+
+    PitchRange() = default;
+    PitchRange(int min, int max)
+        : min(min), max(max) {}
+
+    bool operator ==(const PitchRange& other) const
+    {
+        return min == other.min && max == other.max;
+    }
+
+    bool operator !=(const PitchRange& other) const
+    {
+        return !operator ==(other);
+    }
+};
+
+struct MidiAction
+{
+    QString name;
+    QString description;
+    std::vector<midi::Event> events;
+};
+using MidiActionList = QList<MidiAction>;
+
+using MidiArticulations = QList<Ms::MidiArticulation>;
+
+struct InstrumentGroup
+{
+    QString id;
+    QString name;
+    bool extended = false;
+    int sequenceOrder = 0;
+};
+
+using InstrumentGroups = QList<InstrumentGroup>;
+
+struct InstrumentGenre
+{
+    QString id;
+    QString name;
+};
+using InstrumentGenres = QList<InstrumentGenre>;
+
+static const QString COMMON_GENRE_ID("common");
+
+struct Instrument
+{
+    QString id;
+    StaffNameList longNames;
+    StaffNameList shortNames;
+    QString name;
+    QString musicXMLid;
+    QString templateId;
+    QString description;
+    int sequenceOrder = 0;
+
+    bool extended = false;
+    int staves = 1;
+
+    QString groupId;
+    QStringList genreIds;
+    QString familyId;
+
+    PitchRange amateurPitchRange;
+    PitchRange professionalPitchRange;
+
+    ClefTypeList clefs[MAX_STAVES];
+    int staffLines[MAX_STAVES] = { 0 };
+    BracketType bracket[MAX_STAVES] = { BracketType::NO_BRACKET };
+    int bracketSpan[MAX_STAVES] = { 0 };
+    int barlineSpan[MAX_STAVES] = { 0 };
+    bool smallStaff[MAX_STAVES] = { false };
+
+    Interval transpose;
+
+    StaffGroup staffGroup = StaffGroup::STANDARD;
+    const StaffTypePreset* staffTypePreset = nullptr;
+
+    bool useDrumset = false;
+    const Drumset* drumset = nullptr;
+
+    StringData stringData;
+
+    bool singleNoteDynamics = false;
+
+    MidiActionList midiActions;
+    QList<MidiArticulation> midiArticulations;
+
+    InstrumentChannelList channels;
+
+    Trait trait;
+
+    bool isValid() const { return !id.isEmpty(); }
+    QString abbreviature() const { return !shortNames.isEmpty() ? shortNames.first().name() : QString(); }
+};
+
+using Instruments = QList<Instrument>;
+
+struct PartInstrument
+{
+    QString partId;
+    Instrument instrument;
+
+    bool isExistingPart = false;
+    bool isSoloist = false;
+};
+
+using PartInstrumentList = QList<PartInstrument>;
+
+struct ScoreOrderGroup
+{
+    QString family;
+    QString section;
+    QString unsorted;
+
+    bool bracket = false;
+    bool showSystemMarkings = false;
+    bool barLineSpan = false;
+    bool thinBracket = false;
+};
+
+using InstrumentOverwrite = Ms::InstrumentOverwrite;
+
+struct ScoreOrder
+{
+    QString id;
+    QString name;
+    QMap<QString, InstrumentOverwrite> instrumentMap;
+    QList<ScoreOrderGroup> groups;
+
+    bool isValid() { return !groups.empty(); }
+};
+
+using ScoreOrders = QList<ScoreOrder>;
+
+struct PartInstrumentListScoreOrder
+{
+    PartInstrumentList instruments;
+    ScoreOrder scoreOrder;
+};
+
+struct InstrumentsMeta
+{
+    Instruments instrumentTemplates;
+    InstrumentGroups groups;
+    InstrumentGenres genres;
+    MidiArticulations articulations;
+    ScoreOrders scoreOrders;
+
+    void clear()
+    {
+        instrumentTemplates.clear();
+        groups.clear();
+        genres.clear();
+        articulations.clear();
+        scoreOrders.clear();
+    }
+};
+
 struct ScoreCreateOptions
 {
     QString title;
@@ -307,8 +492,8 @@ struct ScoreCreateOptions
 
     io::path templatePath;
 
-    instruments::PartInstrumentList parts;
-    instruments::ScoreOrder order;
+    PartInstrumentList parts;
+    ScoreOrder order;
 };
 
 struct SearchCommand
@@ -381,7 +566,7 @@ struct StaffConfig
     bool mergeMatchingRests = false;
     Staff::HideMode hideMode = Staff::HideMode::AUTO;
     NoteHead::Scheme noteheadScheme = NoteHead::Scheme::HEAD_AUTO;
-    Ms::ClefTypeList clefType;
+    ClefTypeList clefTypeList;
 };
 
 struct TransposeOptions
@@ -463,15 +648,16 @@ struct ScoreConfig
 
 inline QString staffTypeToString(StaffType type)
 {
-    return Ms::StaffType::preset(type)->name();
+    const Ms::StaffType* preset = Ms::StaffType::preset(type);
+    return preset ? preset->name() : QString();
 }
 
 inline QList<StaffType> allStaffTypes()
 {
     QList<StaffType> result;
 
-    for (const Ms::StaffType& staffType: Ms::StaffType::presets()) {
-        result << staffType.type();
+    for (const Ms::StaffType& preset: Ms::StaffType::presets()) {
+        result << preset.type();
     }
 
     return result;
@@ -513,5 +699,8 @@ inline bool isFretIndexValid(int fretIndex)
     return 0 <= fretIndex && fretIndex < MAX_FRET;
 }
 }
+
+Q_DECLARE_METATYPE(mu::notation::Instrument)
+Q_DECLARE_METATYPE(mu::notation::ScoreOrder)
 
 #endif // MU_NOTATION_NOTATIONTYPES_H
